renaming binary from qpoint to qtap
