+++ Makefile
@@ -22,11 +22,11 @@ endif
 # 🎯 Project Configuration
 # =============================================================================
 # Project Settings
-PROJECT_NAME ?= qpoint
-BINARY_NAME ?= qpoint
+PROJECT_NAME ?= qtap
+BINARY_NAME ?= qtap
 ORGANIZATION ?= qpoint-io
-DESCRIPTION ?= "🧬 QPoint is a tool for monitoring and analyzing network traffic."
-MAINTAINER ?= "QPoint Team \<hello@qpoint.io\>"
+DESCRIPTION ?= "🧬 Qtap: An eBPF agent that captures pre-encrypted network traffic, providing rich context about egress connections and their originating processes."
+MAINTAINER ?= "Qpoint Team \<hello@qpoint.io\>"
 
 VERSION=$${GIT_VERSION:-$$(git describe --tags --always --dirty)}
 GIT_COMMIT ?= $(shell git rev-parse --short HEAD 2>/dev/null || echo "unknown")
@@ -135,12 +135,12 @@ build: $(BIN_DIR) generate ## Build for the current platform
 .PHONY: run
 run: build ## Run the application
 	@echo $(ROCKET) Running $(PROJECT_NAME)... $(RESET)
-	./bin/$(BINARY_NAME) tap --log-level=debug --log-encoding=console
+	./bin/$(BINARY_NAME) --log-level=debug --log-encoding=console
 
 .PHONY: run-config
 run-config: build ## Run the application with a specific config
 	@echo $(ROCKET) Running $(PROJECT_NAME) with config... $(RESET)
-	./bin/$(BINARY_NAME) tap --log-level=debug --log-encoding=console --config=$$(find ./examples -type f -name "*.yaml" | go tool gum filter --prompt="> " --indicator=">" --placeholder="Select a config file..." --header="Select a config file to run")
+	./bin/$(BINARY_NAME) --log-level=debug --log-encoding=console --config=$$(find ./examples -type f -name "*.yaml" | go tool gum filter --prompt="> " --indicator=">" --placeholder="Select a config file..." --header="Select a config file to run")
 
 .PHONY: generate
 generate: ## Run code generation
